fix build_prompt_input_includes_context_and_user_message
