+++ codex-rs/core/tests/suite/mod.rs
@@ -67,6 +67,7 @@ mod personality;
 mod personality_migration;
 mod plugins;
 mod prompt_caching;
+mod prompt_debug_tests;
 mod quota_exceeded;
 mod realtime_conversation;
 mod remote_env;
